Partial support for the "attr_list" extension.
This extension allow for adding a list of attributes to inline level
elements, and in the future hopefully to block level elements.

An exemple:

    Wikis are **nice**{: #id .class key=value }!

    <p>Wikis are <strong id="id" class="class"
    key="value">nice</strong>!</p>

To enable the extension, do the following:

    marked.setOptions({
      attr_list: true
    });

For a more extensive documentation about the syntax, please refer to
<http://maruku.rubyforge.org/proposal.html#attribute_lists>

Notes: Attribute lists definitions (ALD) is not supported at the moment.